--- config/main.yaml
@@ -1,9 +1,0 @@
-defaults:
-  - process: process1
-  - model: model1
-  - _self_
-
-data:
-  raw: data/copied_locally/sample.csv
-  processed: data/processed/processed.csv
-  final: data/final/final.csv
