--- config/model/model1.yaml
@@ -1,1 +1,0 @@
-name: model1
