--- config/model/model2.yaml
@@ -1,1 +1,0 @@
-name: model2
