--- config/process/process1.yaml
@@ -1,3 +1,0 @@
-use_columns: 
-  - col1
-  - col2
