--- config/process/process2.yaml
@@ -1,4 +1,0 @@
-use_columns: 
-  - col1
-  - col2
-  - col3
