+++ fake_news_detection/config/model/concat_model.yaml
@@ -1,0 +1,13 @@
+name: concat_model_test_10k
+
+train_data: C:\Users\Claire\Documents\UdeM\IFT6759\IFT6759-Fake-news-detection\data\raw\sample\multimodal_train_10k.tsv
+val_data: C:\Users\Claire\Documents\UdeM\IFT6759\IFT6759-Fake-news-detection\data\raw\sample\multimodal_train_1k.tsv
+image_folder: C:\Users\Claire\Documents\UdeM\IFT6759\IFT6759-Fake-news-detection\data\raw\public_image_set
+
+
+batch_size: 16
+num_epochs: 100
+learning_rate: 0.0001
+patience: 5  # Early stopping patience
+
+save_dir: "fake_news_detection/logs/"
