+++ fake_news_detection/src/__init__.py
@@ -1,0 +1,6 @@
+"""Source code of your project"""
+from src.utils import *
+from src.data_fetch import *
+from src.images_fetch import *
+from src.process import *
+from src.train_model import *
