+++ pyproject.toml
@@ -6,13 +6,13 @@ description = ""
 authors = ["maciusyoury15"]
 
 [tool.poetry.dependencies]
-python = "3.12.4"
+python = "3.12.9"
 hydra-core = "^1.1.1"
 
 [tool.poetry.group.dev]
 optional = true
 
-[tool.poetry.group.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 pdoc3 = "^0.10.0"
 pytest = "^6.2.5"
 pre-commit = "^2.17.0"
