--- src/__init__.py
@@ -1,6 +1,0 @@
-"""Source code of your project"""
-from src.utils import *
-from src.data_fetch import *
-from src.images_fetch import *
-from src.process import *
-from src.train_model import *
