--- src/process.py
@@ -1,20 +1,0 @@
-"""
-This is the demo code that uses hydra to access the parameters in under the directory config.
-
-Author: Khuyen Tran
-"""
-
-import hydra
-from omegaconf import DictConfig
-
-
-@hydra.main(config_path="../config", config_name="main", version_base="1.2")
-def process_data(config: DictConfig):
-    """Function to process the data"""
-
-    print(f"Process data using {config.data.raw}")
-    print(f"Columns used: {config.process.use_columns}")
-
-
-if __name__ == "__main__":
-    process_data()
