--- tests/__init__.py
@@ -1,5 +1,0 @@
-# Avoid ModuleNotFoundError
-
-import sys
-
-sys.path.append("./src")
